requirements.txt
groq
requests
beautifulsoup4
streamlit
pandas
python-dotenv